Oversight fix, default values, and sapply->lapply change.
Included calibration for condition 3.2 in  process_dh_aaf.

Derive pc didn't have default arguments.  This is annoying when using that function directly, and doesn't provide users with an example for how to model the arguments, so default arguments were added.

FP factory was using sapply to evaluate as list of functions.  Changed it to lapply (should be faster?  Maybe imperceptibly so.)